pyproject.toml
[tool.poetry]
name = "soroeditor-qt"
version = "0.0.0"
description = "SoroEditor - Parallel Text Editor"
authors = ["joppincal <106348367+joppincal@users.noreply.github.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11, <3.13"
pyyaml = "^6.0.1"
pyside6 = "^6.7.0"
darkdetect = "^0.8.0"
types-pyyaml = "^6.0.12.20240808"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.8.0"

[tool.poetry-dynamic-versioning]
enable = true
format = "{base}"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
